Move all cmds to cogs folder
