+++ src/cogs/admin_commands.py
@@ -1,3 +1,4 @@
+'''Module for admin-only commands in guilds'''
 import disnake
 from disnake.ext import commands
 from src.utils import guild_funcs
@@ -14,16 +15,17 @@ class AdminCommands(commands.Cog):
 
     @admin.sub_command()
     async def clear(self, inter: disnake.CommandInteraction):  # pylint: disable=no-self-use
-        '''/gen clear: Clear all roles with matching names from server'''
+        '''/admin clear: Clear all roles with matching names from server'''
         if inter.author.guild_permissions.administrator:
             await inter.response.defer()
             await guild_funcs.remove_roles_in_guild(inter.guild)
             await inter.edit_original_message(content="All roles cleared")
         else:
-            await inter.response.send_message("You're not an admin, why bother with this")
+            await inter.response.send_message("You're not an admin, why bother with this?")
 
     @admin.sub_command()
     async def show(self, inter: disnake.CommandInteraction):
+        '''/admin show: Show whether the bot considers you to be an admin or not'''
         if inter.author.guild_permissions.administrator:
             await inter.response.send_message("You are an admin")
         else:
